\name{Yeast_hughes}
\alias{Yeast_hughes}
\docType{data}
\title{
Yeast gene expression data
}
\description{
The data is a small subset of the full dataset from Hughes et al. (2000).
}
\usage{data(Yeast_hughes)}
\format{
  A data frame for 10 \emph{S. cerevisae} (which is a common and well-studied species of Yeast) genes (stored in rows) and 300 expression observations (stored in columns). The 10 genes are those positioned 10, 20, 30, 40, 50, 60, 70, 80, 90, and 100 in chromosome I in the original data (see link below).
}
\source{
Functional Discovery via a Compendium of Expression Profiles. Hughes et al., Cell, 2000.

The full data is publicly available at \url{http://noble.gs.washington.edu/proj/microarray/Rosetta/}.
}
\references{
Hughes et al.(2000). Functional Discovery via a Compendium of Expression Profiles. \emph{Cell}, 102(Issue 1), P103-P126.

}
\examples{
data(Yeast_hughes)
}
\keyword{datasets}
